change ple_count variable type

+++ xen/arch/x86/ple_count.c
@@ -1,9 +1,9 @@
-unsigned int ple_count = 0;
+unsigned long long ple_count = 0;
 
-void do_set_ple_count(unsigned int num){
+void do_set_ple_count(unsigned long long num){
 	ple_count = num;
 }
 
-int do_get_ple_count(void){
+unsigned long long do_get_ple_count(void){
 	return	ple_count;
 }
